Enable use of cmake ExternalProject, and vcs tool
sbuild disables network access as per debians policy.

But ROS2 packages need/do access the network
especially when vendoring packages. A yet to be
published option for sbuild will, in the future,
allow this policy to be controlled through configuration.

SBUILD_CONF ENABLE_NETWORK

For now we need to patch it after it has been installed in
prepare, and before we want to use it in build.

See https://github.com/jspricke/ros-deb-builder-action/issues/38